requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
comm==0.1.2
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dm-tree==0.1.8
eccodes==1.5.0
einops==0.6.0
entrypoints==0.4
etils==1.0.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
findlibs==0.0.2
flatbuffers==23.1.4
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.4
gast==0.4.0
gin-config==0.1.1
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
gradient-accumulator==0.3.1
grpcio==1.51.1
gviz-api==1.10.0
h5py==3.7.0
html5lib==1.1
httplib2==0.21.0
huggingface-hub==0.12.0
idna @ file:///croot/idna_1666125576474/work
immutabledict==2.2.3
importlib-resources==5.10.2
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jax==0.4.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.0.6
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.12
keopscore==2.1.1
keras==2.11.0
keras-nightly==2.12.0.dev2023012708
keras-nlp==0.4.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.4
multitasking==0.0.11
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
portalocker==2.7.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycocotools==2.0.6
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pykeops==2.1.1
Pympler==1.0.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==8.0.0
pytickersymbols==1.13.0
pytz==2022.7
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
sacrebleu==2.3.1
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.0
sentencepiece==0.1.97
seqeval==1.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
svgwrite==1.4.3
tabulate==0.9.0
tb-nightly==2.12.0a20230127
tensorboard==2.11.0
tensorboard-data-server==0.7.0
tensorboard-plugin-profile==2.11.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.2
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflow-text-nightly==2.12.0.dev20230127
tensorrt==8.5.2.2
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tf-estimator-nightly==2.12.0.dev2023012709
tf-models-official==2.11.3
tf-nightly==2.12.0.dev20230124
tf-slim==1.1.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
toolz @ file:///croot/toolz_1667464077321/work
torch-tb-profiler==0.4.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.8.1
transformers==4.26.1
typeguard==2.13.3
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1673575502006/work
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xarray==2022.12.0
yfinance==0.2.11
zipp==3.13.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
